Refactor MRepo index loading
